feature: more strict checking against #(args) passed to the geth.ps1 script

+++ geth.ps1
@@ -1,1 +1,7 @@
+
+if ($args.Count -eq 0) {
+  echo 'too few arguments'
+  exit
+}
+
 docker run --rm ethereum/client-go:alltools-v1.8.23 geth $args
